[BE] feat: 투두리스트

+++ src/main/java/com/example/hackaton/repository/EmotionalScoreRepository.java
@@ -8,5 +8,5 @@ import java.util.List;
 
 @Repository
 public interface EmotionalScoreRepository extends JpaRepository<EmotionalScore, Long> {
-    List<EmotionalScore> findByGoogleId(int googleId);
+    List<EmotionalScore> findByGoogleId(long googleId);
 }
